fix: changed datatype of bitstring to list of floats

+++ Project_1/sga/sga.py
@@ -3,7 +3,7 @@ class Individual:
 
     def __init__(
         self,
-        bitstring: list[int],
+        bitstring: list[float],
         phenotype: str,
         fitness: float = 0,
         parents: "list[Individual]" = None,
